Added test cases; refactored to allow testing with test random generator; changed type of totals to long to avoid overflow

+++ src/main/java/tech/blackfall/dicedist/simulation/domain/SimulationPartialResult.java
@@ -5,6 +5,6 @@ import lombok.Value;
 @Value
 public class SimulationPartialResult {
 
-  int totalValue;
+  long totalValue;
   int numberOfOccurrences;
 }
